Move icons out of /app folder
